added nav bar and more styling
Added a navigation bar with a color mode button and a github button.

Added styling to the category buttons to show which category is currently visible.

Added text that says "Start adding tasks!" when there are no tasks to be displayed.

Fixed a bug that occured when adding a new task.